LayoutTests/platform/mac-wk2/platform/mac/accessibility/selected-rows-table-expected.txt

a


b


c
This tests that the selected rows of a table are exposed correctly and that when they change, a notification is sent.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
FAIL axTree.selectedRowAtIndex(0).isEqual(axTree.rowAtIndex(0)) should be true. Threw exception TypeError: 'undefined' is not a function (evaluating 'axTree.rowAtIndex(0)')
FAIL axTree.selectedRowAtIndex(1).isEqual(axTree.rowAtIndex(2)) should be true. Threw exception TypeError: 'undefined' is not a function (evaluating 'axTree.rowAtIndex(2)')
PASS successfullyParsed is true

TEST COMPLETE
Notification: AXSelectedRowsChanged
